package/ola: ftdidmx needs gcc >= 4.9
Fix the following build failure raised since re-introduction of the
package in commit 770e0f66fa35ea9ded914c3880c5457066d0c8c3:

Makefile:575: *** libusb is in the dependency chain of libftdi1 that has added it to its _DEPENDENCIES variable without selecting it or depending on it from Config.in.  Stop.

Fixes:
 - http://autobuild.buildroot.org/results/f65baf4dd4c9a313745cdb9aefad2eae7a295469

Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>