package/openlayers: bump to version 7.1.0
OpenLayers changed its release format again, only two weeks after the
previous change.  The good news is, though, that they now ship a
single release archive that includes the compiled JavaScript code
together with the source code.  Furthermore, they ship both code with
a single license, which file is part of the archive.

Hence, we still do not need to rely on host-nodejs for this package,
at least not for this version.

Signed-off-by: Thomas Claveirole <thomas.claveirole@green-communications.fr>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>