package/mesa3d-demos: bump version to 8.5.0
The package build system changed from autotools to meson. Autotools
was deprecated in commit [1], which is included in this 8.5.0
version. It was definitevely removed soon after, in commit [2].

The package patch file making libgl optional was fully rewritten
for meson.

This patch also updates the old ftp _SITE url to the new https one,
published in the project README.rst file. See [3].

This patch also removes openvg and freetype2 dependencies, as their
support was removed in commit [4]. The 'gbm' option is removed, and
the option 'libdrm' is added.

The md5 and sha1 entries in hash files are also removed, to keep only
sha256 and sha512 from the release announce email [5].

For change log, see the release announce [5].

[1] https://gitlab.freedesktop.org/mesa/demos/-/commit/2df8e9710834c3b42ce5cfb67b7662c110ab904e
[2] https://gitlab.freedesktop.org/mesa/demos/-/commit/abfead37587f1f2e124bdfcde755b6e90421b538
[3] https://gitlab.freedesktop.org/mesa/demos/-/blob/mesa-demos-8.5.0/README.rst
[4] https://gitlab.freedesktop.org/mesa/demos/-/commit/e6d0d9a8682c544870fd41983ea2e0484b9767b1
[5] https://lists.freedesktop.org/archives/mesa-announce/2022-May/000677.html

Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>