configs/visionfive_defconfig: new defconfig
Add a defconfig for the Starfive VisionFive board, a board built around the
Starfive JH7100 RISC-V 64bit SoC (same as Beaglev).

This board comes with functional lowlevel and U-Boot bootloaders in SPI
flash.  The defconfig reuses these and only builds a (6.0 based) kernel and
rootfs.

The factory shipped U-Boot is hard coded to look at MMC partition 3 and
misses some variables, so we provide a uEnv.txt to fix that up, based on
what is done in provided Fedora image.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>